BackEnd/requirements.txt
absl-py==2.1.0
alembic==1.14.1
aniso8601==10.0.0
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.42.2
colorama==0.4.6
cryptography==44.0.0
dropbox==12.0.2
fastapi==0.115.11
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Migrate==4.1.0
Flask-MySQLdb==2.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.4.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
mysql-connector==2.2.9
mysql-connector-python==9.1.0
mysql-connector-python-rf==2.2.2
mysqlclient==2.2.7
numpy==2.2.2
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
ply==3.11
proto-plus==1.26.0
protobuf==5.29.3
psycopg==3.2.4
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.6.1
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
starlette==0.46.0
stone==3.3.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.2
Werkzeug==3.1.3
WTForms==3.2.1
